@@ -30,7 +30,7 @@
   "engineStrict": true,
   "engines": {
     "npm": ">=3.0.0",
-    "node": ">=16"
+    "node": ">=20"
   },
   "eslintConfig": {
     "extends": [
@@ -85,9 +85,9 @@
     "memoize-one": "^5.1.1",
     "openseadragon": "^4.1.1",
     "pako": "^2.0.3",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-async": "^10.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-flip-toolkit": "^7.0.12",
     "react-icons": "^4.2.0",
     "react-redux": "^7.2.0",
@@ -117,8 +117,8 @@
     "@babel/register": "^7.13.16",
     "@babel/runtime": "^7.13.16",
     "@blueprintjs/eslint-plugin": "^0.3.0",
-    "@chromatic-com/playwright": "^0.5.3",
-    "@playwright/test": "^1.40.1",
+    "@chromatic-com/playwright": "^0.6.18",
+    "@playwright/test": "^1.46.1",
     "@sentry/webpack-plugin": "^1.15.0",
     "@storybook/addon-essentials": "^7.6.12",
     "@storybook/addon-interactions": "^7.6.12",
@@ -147,8 +147,8 @@
     "@types/node": "^20.10.6",
     "@types/openseadragon": "^3.0.10",
     "@types/pako": "^1.0.2",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.3.4",
+    "@types/react-dom": "^18.3.0",
     "@types/react-redux": "^7.1.18",
     "@types/sha1": "^1.1.3",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
@@ -159,7 +159,7 @@
     "browserslist-useragent-regexp": "^3.0.2",
     "chalk": "^4.1.0",
     "cheerio": "^1.0.0-rc.6",
-    "chromatic": "^10.7.1",
+    "chromatic": "^11.7.1",
     "chromatic-playwright": "^0.4.1",
     "clean-css": "^5.1.2",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
@@ -198,12 +198,12 @@
     "prettier": "^2.0.5",
     "rimraf": "^3.0.2",
     "serve-favicon": "^2.5.0",
-    "storybook": "^7.6.12",
+    "storybook": "^8.2.9",
     "terser-webpack-plugin": "^5.1.1",
     "typescript": "^4.3.5",
     "webpack": "^5.34.0",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-middleware": "^4.1.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-middleware": "^5.3.4",
     "webpack-merge": "^5.0.9"
   },
   "babel": {
